CMOSD-2187 Importing old assets/attachments from S3/localdisk to MongoDB

+++ app/controllers/rails_admin/main_controller.rb
@@ -32,33 +32,52 @@ module RailsAdmin
       super if @scope_adapter
     end
 
-    def import_attachments
-      s3 = Aws::S3::Resource.new(
+    def import_attachments      
+      if CaseCenter::Config::Reader.get("s3_assets_bucket") 
+        using_S3 = true   
+        s3 = Aws::S3::Resource.new(
         access_key_id: CaseCenter::Config::Reader.get("aws_access_key_id"), 
         secret_access_key: CaseCenter::Config::Reader.get("aws_secret_key"), 
-        region: CaseCenter::Config::Reader.get("s3_region"))
-      bucketName = CaseCenter::Config::Reader.get("s3_assets_bucket")
-      bucket = s3.bucket(bucketName)
-      compId = @company.id
-      pref = "ckeditor_assets/attachments/"+@company.key
-      bucketObjects = bucket.objects(prefix: "#{pref}").collect(&:key)
-      result = ActiveRecord::Base.connection.exec_query("SELECT data_file_name, assetable_type from ckeditor_assets where type = 'Ckeditor::AttachmentFile' and assetable_id = #{compId}")
-      mySqlNames = result.to_a.map{|myItem|myItem["data_file_name"]}
-      bucketObjects.each do |b|
-        splitB = b.split('/')[-1]
-        normalizedB = splitB.downcase.tr(" ", "_")
-        if mySqlNames.include?(normalizedB)
-          result.rows.each do |row|
-            if row[0] == normalizedB
-              @recordId = row[1]
-              @records = @application.get_mongoid_class.find_by(id: @recordId) rescue nil
-            end
+        region: CaseCenter::Config::Reader.get("s3_region"))        
+        bucketName = CaseCenter::Config::Reader.get("s3_assets_bucket")
+        bucket = s3.bucket(bucketName)
+        
+        pref = "ckeditor_assets/attachments/bestegg" # +@company.key
+        importObjects = bucket.objects(prefix: "#{pref}").collect(&:key)
+      elsif CaseCenter::Config::Reader.get("local_assets_path")
+        using_S3 = false
+        # local_path = CaseCenter::Config::Reader.get("local_assets_path") + "/ckeditor_assets/attachments/" + @company.key
+        local_path = "/CaseCenterAPP/old_stuff/ckeditor_assets/attachments/bestegg"
+        files = Dir["#{local_path}/**/*"]
+        importObjects = files.each.map { |f| f if File.file?(f) }.compact        
+      end
+           
+      result = ActiveRecord::Base.connection.exec_query("SELECT unique_id, assetable_type from ckeditor_assets where type = 'Ckeditor::AttachmentFile' and assetable_id = #{@company.id}")
+
+      importObjects.each do |b|
+        object_uid = b.split('/')[-2]
+        import_filename = b.split('/')[-1].downcase.tr(" ", "_")
+        result.rows.each do |row|
+          record = nil
+          if b.include? row[0] # match by unique_id
+            record_id = row[1]
+            record = @application.get_mongoid_class.find_by(id: record_id) rescue nil
           end
-          if @records
-            fileDownloaded = s3.bucket(bucketName).object("#{b}")
-            fileDownloaded.get(response_target: "#{Rails.root}/tmp/#{normalizedB}")
+          if record
+            if using_S3
+              fileDownloaded = s3.bucket(bucketName).object("#{b}")
+              fileDownloaded.get(response_target: "#{Rails.root}/tmp/#{import_filename}")
+              attFile = File.open("#{Rails.root}/tmp/#{import_filename}")
+              if attFile.size == 0
+                File.delete(attFile.path) 
+                next
+              end
+            else
+              attFile = File.open(b)
+              next if attFile.size == 0
+            end            
+
             @newAsset = Attachment.new
-            attFile = File.open("#{Rails.root}/tmp/#{normalizedB}")
             currentFileType = Terrapin::CommandLine.new('file', '-b --mime-type :file').run(file: attFile.path).strip
             if(CaseCenter::Config::Reader.get('mongodb_attachment_database'))
               Mongoid.override_client(:attachDb)
@@ -74,23 +93,26 @@ module RailsAdmin
             encData = cipher.update(File.read(encFile))
             encData << cipher.final
             #End Encryption
-            File.open(encFile, 'wb') do |f|
-              f.write(encData)
-            end
-            grid_file = grid_fs.put(attFile.path)
+
+            File.open("#{Rails.root}/tmp/#{import_filename}_encrypted", "wb") {|f| f.write(encData) }
+            file_encrypted = File.open("#{Rails.root}/tmp/#{import_filename}_encrypted")
+            grid_file = grid_fs.put(file_encrypted.path)
             encrypted_aes = Base64.encode64(public_key.public_encrypt(key))
             @newAsset.aes_key = encrypted_aes
             @newAsset.data = grid_file.id #Attachment.data is equal to the BSON::ObjectId of the GridFs file.
-            @newAsset.company_id = @records.system.company_id
+            @newAsset.company_id = record.system.company_id
 
-            @newAsset.record_id = @recordId
-            @newAsset.data_file_name = normalizedB
+            @newAsset.record_id = record_id
+            @newAsset.data_file_name = import_filename
             @newAsset.user = current_user.email
 
             @newAsset.data_file_size = File.size(attFile.path).to_i
             @newAsset.updated_at = Time.now.strftime("%a %b %e %Y, %k:%M:%S")
             attFile.close
-          
+            if using_S3
+              File.delete(attFile.path)
+            end
+            File.delete(file_encrypted.path)
             if(CaseCenter::Config::Reader.get('mongodb_attachment_database'))
               Mongoid.override_client(:default)
             end
@@ -103,12 +125,12 @@ module RailsAdmin
             @newAsset.destroy if @newAsset
             raise e
           ensure
-            count = Attachment.where(:record_id => @recordId).size
-            @records.system.update_attributes(
+            count = Attachment.where(:record_id => record_id).size
+            record.system.update_attributes(
               attachments_count: count,
               edited_by: current_user.email,
               edited_by_role: current_user.roles.map(&:name)
-            ) if @records
+            ) if record
             Mongoid.override_client(:default)
           end
         end
@@ -117,38 +139,52 @@ module RailsAdmin
       redirect_to "/admin/Company/"+@company.id.to_s+"/edit?locale=en"
     end
 
-    def import_assets
-      s3 = Aws::S3::Resource.new(
-        access_key_id: CaseCenter::Config::Reader.get("aws_access_key_id"), 
-        secret_access_key: CaseCenter::Config::Reader.get("aws_secret_key"), 
-        region: CaseCenter::Config::Reader.get("s3_region"))
-      bucketName = CaseCenter::Config::Reader.get("s3_assets_bucket")
-      bucket = s3.bucket(bucketName)
-      compId = @company.id
-      pref = "ckeditor_assets/pictures/" + @company.key
-      bucketObjects = bucket.objects(prefix: "#{pref}").collect(&:key)
+    def import_assets      
 
-      result = ActiveRecord::Base.connection.exec_query("SELECT unique_id from ckeditor_assets where type = 'Ckeditor::Picture' and assetable_id = #{compId}")
+      if CaseCenter::Config::Reader.get("s3_assets_bucket")
+        using_S3 = true
+        s3 = Aws::S3::Resource.new(
+          access_key_id: CaseCenter::Config::Reader.get("aws_access_key_id"), 
+          secret_access_key: CaseCenter::Config::Reader.get("aws_secret_key"), 
+          region: CaseCenter::Config::Reader.get("s3_region"))
+        bucketName = CaseCenter::Config::Reader.get("s3_assets_bucket")
+        bucket = s3.bucket(bucketName)
+        pref = "ckeditor_assets/pictures/" + @company.key
+        importObjects = bucket.objects(prefix: "#{pref}").collect(&:key)
+      elsif CaseCenter::Config::Reader.get("local_assets_path")
+        using_S3 = false
+        local_path = CaseCenter::Config::Reader.get("local_assets_path") + "/ckeditor_assets/pictures/" + @company.key
+        files = Dir["#{local_path}/**/*"]            
+        importObjects = files.each.map { |f| f if File.file?(f) and f.include? "content_" }.compact 
+      end
+
+      result = ActiveRecord::Base.connection.exec_query("SELECT unique_id from ckeditor_assets where type = 'Ckeditor::Picture' and assetable_id = #{@company.id}")
       result.rows.each do |row|
-        bucketObjects.each do |b|
+        importObjects.each do |b|
           if( b.include?(row[0]) and  b.split('/')[-1].starts_with? "content_")          
             splitB = b.split('/')[-1]
             splitB[0..7]  = ""
-            normalizedB = splitB.downcase.tr(" ", "_") 
+            normalizedB = splitB.downcase.tr(" ", "_")
 
             picture_asset = PictureAsset.new
 
             if PictureAsset.where(company_id: @company.id,  data_file_name: normalizedB ).size > 0 
               if normalizedB.include?(".")
-                normalizedB = normalizedB.match(/.*(?=\.)/)[0] + "_1." + normalizedB.match(/\.[^.]+$/)[0]
+                normalizedB = normalizedB.match(/.*(?=\.)/)[0] + "_1" + normalizedB.match(/\.[^.]+$/)[0]
               else
                 normalizedB = normalizedB + "_1"
               end
             end
 
-            fileDownloaded = s3.bucket(bucketName).object("#{b}")
-            fileDownloaded.get(response_target: "#{Rails.root}/tmp/#{normalizedB}")
-            file = File.open("#{Rails.root}/tmp/#{normalizedB}")
+            if using_S3
+              fileDownloaded = s3.bucket(bucketName).object("#{b}")
+              fileDownloaded.get(response_target: "#{Rails.root}/tmp/#{normalizedB}")
+              file = File.open("#{Rails.root}/tmp/#{normalizedB}")
+            else
+              file = File.open(b)
+            end
+
+            next if file.size == 0
 
             picture_asset.data_file_name = normalizedB
             picture_asset.data_content_type = MIME::Types.type_for("#{Rails.root}/tmp/#{normalizedB}").first.content_type
@@ -175,21 +211,25 @@ module RailsAdmin
                   f.write(encThumbData)
                 end
                 encData = cipher.update(File.read(file))
-                # encData << cipher.final
-                File.open(file, 'wb') do |f|
-                  f.write(encData)
-                end
+                encData << cipher.final
+
+                File.open("#{Rails.root}/tmp/#{normalizedB}_encrypted", "wb") {|f| f.write(encData) }
+                file_encrypted = File.open("#{Rails.root}/tmp/#{normalizedB}_encrypted")
+
                 encrypted_aes = Base64.encode64(public_key.public_encrypt(key))
                 picture_asset.aes_key = encrypted_aes
                 #End of Encryption
-                grid_file = grid_fs.put(file.path)
+                grid_file = grid_fs.put(file_encrypted.path)
                 picture_asset.data_file_size = File.size(file).to_i
                 picture_asset.company_id = @company.id.to_i
                 picture_asset.image_id = grid_file.id
                 grid_thumb_file = grid_fs.put(thumbFile.path)
                 picture_asset.thumb_image_id = grid_thumb_file.id
+                if using_S3
+                  File.delete(file.path)
+                end
                 File.delete(thumbFile.path)
-                File.delete(file.path)
+                File.delete(file_encrypted.path)
                 if(CaseCenter::Config::Reader.get('mongodb_attachment_database'))
                   Mongoid.override_client(:default)
                 end
